openssl-0.9.7l.tar.gz f0e4136639b10cbd1227c4f7350ff7ad406e575d
$ gpg -v openssl-0.9.7l.tar.gz.asc
gpg: armor header: Version: GnuPG v1.4.2.2 (GNU/Linux)
gpg: assuming signed data in `../openssl-0.9.7l.tar.gz'
gpg: Signature made Thu 28 Sep 2006 13:03:15 BST using RSA key ID 49A563D9
gpg: using PGP trust model
gpg: Good signature from "Mark Cox <mjc@redhat.com>"
gpg:                 aka "Mark Cox <mark@awe.com>"
gpg:                 aka "Mark Cox <mjc@apache.org>"
gpg: WARNING: This key is not certified with sufficiently trusted signatures!
gpg:          It is not certain that the signature belongs to the owner.
Primary key fingerprint: 7B 79 19 FA 71 6B 87 25  0E 77 21 E5 52 D9 83 BF
gpg: binary signature, digest algorithm SHA1

+++ apps/rand.c
@@ -205,7 +205,7 @@ int MAIN(int argc, char **argv)
 		int chunk;
 
 		chunk = num;
-		if (chunk > (int)sizeof(buf))
+		if (chunk > sizeof buf)
 			chunk = sizeof buf;
 		r = RAND_bytes(buf, chunk);
 		if (r <= 0)
